+++ pkg/config/loaders.go
@@ -10,20 +10,38 @@ import (
 	"github.com/apolloconfig/agollo/v4/env/config"
 	"github.com/bytedance/sonic"
 	"github.com/redis/go-redis/v9"
+	"github.com/spf13/cast"
+	"github.com/spf13/viper"
 	"golang.org/x/exp/mmap"
 
 	"github.com/sunhailin-Leo/data-pipeline-go/pkg/logger"
 	"github.com/sunhailin-Leo/data-pipeline-go/pkg/utils"
 )
 
-//go:embed config.json
-var defaultConfigFile []byte
-
 type TunnelConfigLoader struct {
 	config *TunnelConfig
 
 	apolloClient agollo.Client
 	redisClient  redis.UniversalClient
+}
+
+func (c *TunnelConfigLoader) bindAllConfigEnv() {
+	// Load from source
+	_ = viper.BindEnv(utils.ConfigFromSourceName)
+	// Local
+	_ = viper.BindEnv(utils.ConfigFromLocalPathEnvName)
+	// Apollo
+	_ = viper.BindEnv(utils.ConfigFromApolloEnvHost)
+	_ = viper.BindEnv(utils.ConfigFromApolloEnvAppId)
+	_ = viper.BindEnv(utils.ConfigFromApolloEnvNamespace)
+	_ = viper.BindEnv(utils.ConfigFromApolloEnvClusterKey)
+	_ = viper.BindEnv(utils.ConfigFromApolloEnvConfigKey)
+	// Redis
+	_ = viper.BindEnv(utils.ConfigFromRedisEnvHost)
+	_ = viper.BindEnv(utils.ConfigFromRedisEnvUsername)
+	_ = viper.BindEnv(utils.ConfigFromRedisEnvPassword)
+	_ = viper.BindEnv(utils.ConfigFromRedisEnvDBNum)
+	_ = viper.BindEnv(utils.ConfigFromRedisEnvConfigKey)
 }
 
 func (c *TunnelConfigLoader) IsConfigLoaded() bool {
@@ -39,15 +57,9 @@ func (c *TunnelConfigLoader) loadFile(data []byte) *TunnelConfig {
 	return &defaultConfig
 }
 
-// LoadDefaultConfig load default config
-func (c *TunnelConfigLoader) loadDefaultConfig() {
-	TunnelCfg = c.loadFile(defaultConfigFile)
-	c.config = TunnelCfg
-}
-
-func (c *TunnelConfigLoader) loadFromLocal() {
+func (c *TunnelConfigLoader) loadFromLocal(path string) {
 	// use mmap to load file
-	reader, readerErr := mmap.Open(c.config.Config.Local.Path)
+	reader, readerErr := mmap.Open(path)
 	if readerErr != nil {
 		logger.Logger.Fatal("Project configuration file read failed! Error reason: " + readerErr.Error())
 		os.Exit(1)
@@ -62,23 +74,59 @@ func (c *TunnelConfigLoader) loadFromLocal() {
 }
 
 func (c *TunnelConfigLoader) loadFromApollo() {
-	apolloHost := c.config.Config.Apollo.Host
-	if apolloHost == "" {
-		logger.Logger.Fatal(utils.LogServiceName + "Load Apollo config error, reason: host is empty")
+	// Host
+	apolloHostObj := viper.Get(utils.ConfigFromApolloEnvHost)
+	if apolloHostObj == nil {
+		logger.Logger.Fatal(utils.LogServiceName + "Load Apollo config error, reason: APOLLO_HOST is empty")
 		os.Exit(1)
 	}
-	apolloAppId := c.config.Config.Apollo.AppID
-	if apolloAppId == "" {
-		apolloAppId = utils.ServiceName
-		logger.Logger.Warn(utils.LogServiceName + "Load Apollo config error, reason: appId is empty, use default appId instead")
-	}
-	logger.Logger.Info(utils.LogServiceName + "Current using Apollo host: " + apolloHost)
+	apolloHost := apolloHostObj.(string)
 
+	// AppID
+	var apolloAppId string
+	apolloAppIdObj := viper.Get(utils.ConfigFromApolloEnvAppId)
+	if apolloAppIdObj == nil {
+		apolloAppId = utils.ServiceName
+		logger.Logger.Warn(utils.LogServiceName + "Load Apollo config error, reason: APOLLO_APP_ID is empty, use default AppID to instead")
+	} else {
+		apolloAppId = apolloAppIdObj.(string)
+	}
+
+	// Namespace
+	var apolloNamespace string
+	apolloNamespaceObj := viper.Get(utils.ConfigFromApolloEnvNamespace)
+	if apolloNamespaceObj == nil {
+		apolloNamespace = "application"
+		logger.Logger.Warn(utils.LogServiceName + "Load Apollo config error, reason: APOLLO_NAMESPACE is empty, use default NamespaceName to instead")
+	} else {
+		apolloNamespace = apolloNamespaceObj.(string)
+	}
+
+	// Cluster key
+	var apolloClusterKey string
+	apolloClusterKeyObj := viper.Get(utils.ConfigFromApolloEnvClusterKey)
+	if apolloClusterKeyObj == nil {
+		apolloClusterKey = "default"
+		logger.Logger.Warn(utils.LogServiceName + "Load Apollo config error, reason: APOLLO_CLUSTER_KEY is empty, use default NamespaceName to instead")
+	} else {
+		apolloClusterKey = apolloClusterKeyObj.(string)
+	}
+
+	// Config Key
+	var apolloConfigKey string
+	apolloConfigKeyObj := viper.Get(utils.ConfigFromApolloEnvConfigKey)
+	if apolloConfigKeyObj == nil {
+		logger.Logger.Fatal(utils.LogServiceName + "Load Apollo config error, reason: APOLLO_CONFIG_KEY is empty")
+		os.Exit(1)
+	}
+	apolloConfigKey = apolloConfigKeyObj.(string)
+
+	// apollo config
 	cfg := &config.AppConfig{
 		AppID:             apolloAppId,
-		Cluster:           c.config.Config.Apollo.ClusterKey,
+		Cluster:           apolloClusterKey,
 		IP:                apolloHost,
-		NamespaceName:     c.config.Config.Apollo.Namespace,
+		NamespaceName:     apolloNamespace,
 		SyncServerTimeout: 3, // hard code 3 secs
 		MustStart:         true,
 	}
@@ -91,7 +139,7 @@ func (c *TunnelConfigLoader) loadFromApollo() {
 	c.apolloClient = client
 
 	// get config from apollo configuration
-	cache := c.apolloClient.GetConfigCache(c.config.Config.Apollo.Namespace)
+	cache := c.apolloClient.GetConfigCache(apolloNamespace)
 	count := 0
 	cache.Range(func(key, value interface{}) bool {
 		count++
@@ -100,7 +148,7 @@ func (c *TunnelConfigLoader) loadFromApollo() {
 	if count < 1 {
 		logger.Logger.Warn(utils.LogServiceName + "Apollo config is empty! downgrade to using default config")
 	} else {
-		value, loadCacheErr := cache.Get(c.config.Config.Apollo.ConfigKey)
+		value, loadCacheErr := cache.Get(apolloConfigKey)
 		if loadCacheErr != nil {
 			logger.Logger.Error(utils.LogServiceName + "Apollo local config is empty! Error reason: " + loadCacheErr.Error())
 		}
@@ -111,11 +159,53 @@ func (c *TunnelConfigLoader) loadFromApollo() {
 }
 
 func (c *TunnelConfigLoader) loadFromRedis() {
+	// Redis Host
+	redisHostObj := viper.Get(utils.ConfigFromRedisEnvHost)
+	if redisHostObj == nil {
+		logger.Logger.Fatal(utils.LogServiceName + "Load Redis config error, reason: REDIS_HOST is empty")
+		os.Exit(1)
+	}
+	redisHosts := redisHostObj.(string)
+
+	// Redis Username (Redis 6.0+)
+	var redisUsername string
+	redisUsernameObj := viper.Get(utils.ConfigFromRedisEnvUsername)
+	if redisUsernameObj != nil {
+		redisUsername = redisUsernameObj.(string)
+	}
+
+	// Redis Password
+	var redisPassword string
+	redisPasswordObj := viper.Get(utils.ConfigFromRedisEnvPassword)
+	if redisPasswordObj != nil {
+		redisPassword = redisPasswordObj.(string)
+	}
+
+	// Redis DB Num
+	var redisDBNum int
+	redisDBNumObj := viper.Get(utils.ConfigFromRedisEnvDBNum)
+	if redisDBNumObj != nil {
+		redisDBNum = cast.ToInt(redisDBNumObj)
+	} else {
+		// maybe is cluster mode
+		redisDBNum = 0
+	}
+
+	// Redis Config Key
+	var redisConfigKey string
+	redisConfigKeyObj := viper.Get(utils.ConfigFromRedisEnvConfigKey)
+	if redisConfigKeyObj == nil {
+		logger.Logger.Fatal(utils.LogServiceName + "Load Redis config error, reason: REDIS_CONFIG_KEY is empty")
+		os.Exit(1)
+	}
+	redisConfigKey = redisConfigKeyObj.(string)
+
+	// Redis Client
 	client := redis.NewUniversalClient(&redis.UniversalOptions{
-		Addrs:    strings.Split(c.config.Config.Redis.Host, ","),
-		Username: c.config.Config.Redis.Username,
-		Password: c.config.Config.Redis.Password,
-		DB:       c.config.Config.Redis.DB,
+		Addrs:    strings.Split(redisHosts, ","),
+		Username: redisUsername,
+		Password: redisPassword,
+		DB:       redisDBNum,
 	})
 	if pingErr := client.Ping(context.Background()).Err(); pingErr != nil {
 		logger.Logger.Error(utils.LogServiceName + "Failed to connect Redis! Reason for exception: " + pingErr.Error())
@@ -123,7 +213,7 @@ func (c *TunnelConfigLoader) loadFromRedis() {
 	}
 	c.redisClient = client
 
-	rConfig, getErr := c.redisClient.Get(context.Background(), c.config.Config.Redis.ConfigKey).Result()
+	rConfig, getErr := c.redisClient.Get(context.Background(), redisConfigKey).Result()
 	if getErr != nil {
 		logger.Logger.Fatal(utils.LogServiceName + "load Redis config error,reason: " + getErr.Error())
 		return
@@ -135,17 +225,34 @@ func (c *TunnelConfigLoader) loadFromRedis() {
 
 // Load load config api
 func (c *TunnelConfigLoader) Load() {
-	c.loadDefaultConfig()
+	// bind viper setting
+	c.bindAllConfigEnv()
+	// Check source env
+	loadSrcObj := viper.Get(utils.ConfigFromSourceName)
+	if loadSrcObj == nil {
+		logger.Logger.Fatal(utils.LogServiceName + "config load error! CONFIG_SRC env is not set")
+		os.Exit(1)
+	}
 
-	if c.config.Config != nil {
-		switch c.config.Config.From {
-		case utils.ConfigFromLocalTagName:
-			c.loadFromLocal()
-		case utils.ConfigFromApolloTagName:
-			c.loadFromApollo()
-		case utils.ConfigFromRedisTagName:
-			c.loadFromRedis()
+	switch loadSrcObj.(string) {
+	case utils.ConfigFromLocalTagName:
+		localPathObj := viper.Get(utils.ConfigFromLocalPathEnvName)
+		if localPathObj == nil {
+			logger.Logger.Fatal(utils.LogServiceName + "config load error! LOCAL_PATH env is not set!")
+			os.Exit(1)
 		}
+		c.loadFromLocal(localPathObj.(string))
+	case utils.ConfigFromApolloTagName:
+		c.loadFromApollo()
+	case utils.ConfigFromRedisTagName:
+		c.loadFromRedis()
+	default:
+		logger.Logger.Fatal(utils.LogServiceName + "config load error! CONFIG_SRC env is unsupported!")
+		os.Exit(1)
+	}
+
+	if TunnelCfg != nil {
+		c.config = TunnelCfg
 	}
 }